refactor(core): all

+++ src/main/java/com/nopay/nopayapi/repository/invoice/InvoiceItemRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.nopay.nopayapi.entity.invoice.InvoiceItem;
 
 @Repository
-public interface InvoiceItemRepository extends JpaRepository<InvoiceItem, Long> {  
+public interface InvoiceItemRepository extends JpaRepository<InvoiceItem, Integer> {
 }
