+++ src/main/java/com/nopay/nopayapi/repository/products/ColourRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.nopay.nopayapi.entity.products.Colour;
 
-
 @Repository
-public interface  ColourRepository extends JpaRepository<Colour, Long> {
+public interface ColourRepository extends JpaRepository<Colour, Integer> {
     ///
 }
