+++ src/main/java/com/nopay/nopayapi/repository/products/MaterialRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.nopay.nopayapi.entity.products.Material;
 
-
 @Repository
-public interface  MaterialRepository extends JpaRepository<Material, Long> {
+public interface MaterialRepository extends JpaRepository<Material, Integer> {
     ///
 }
